feat(kikan-tauri): V10a — move try_bind_ephemeral_loopback from services/api [#507]
The ephemeral-loopback bind strategy ("127.0.0.1:0", OS-assigned port) is
specific to the Tauri desktop webview shell — it belongs in the Tauri adapter
crate, not in services/api (I2 adapter-boundary hygiene).

- Create crates/kikan-tauri/src/net.rs with try_bind_ephemeral_loopback.
- Expose via kikan_tauri::try_bind_ephemeral_loopback re-export.
- Delete the fn from services/api/src/lib.rs; leave a pointer comment.
- Do NOT re-export from services/api: that would drag `tauri` into mokumo-server
  via the mokumo-api dep and violate I3 (headless binary has no Tauri deps).
- Update apps/mokumo-desktop/src/lib.rs callers to import from kikan_tauri.
- Move the BDD ephemeral-bind step import to kikan_tauri; add kikan-tauri as a
  dev-dep of services/api (dev-deps do not leak into mokumo-server).
- Add tokio + tracing deps to kikan-tauri.

Quality gates green: cargo fmt, cargo check --workspace --all-targets
(excluding mokumo-desktop which has a pre-existing build-script demo.db env
issue — desktop lib itself compiles), cargo test --workspace --lib (478
passed), and all four invariant checks (I1/I2/I3/I4).

V10b (rewire desktop to compose Engine::build_router + axum::serve directly,
replacing mokumo_api::build_app_with_shutdown) deferred — substantial change
that touches restart-loop wiring and mDNS state composition; per plan, tracked
as a separate commit.
