+++ crates/kikan-tauri/src/lib.rs
@@ -1,2 +1,6 @@
-// Stage 0 stub — Tauri IPC adapter.
+// kikan-tauri — Tauri IPC adapter and Tauri-shell-specific helpers.
 // #[tauri::command] wrappers over kikan::control_plane handlers migrate here in Stage 4.
+
+pub mod net;
+
+pub use net::try_bind_ephemeral_loopback;
