+++ services/api/src/lib.rs
@@ -667,21 +667,12 @@ pub async fn try_bind(
     ))
 }
 
-/// Bind an ephemeral loopback port for the desktop shell.
-///
-/// Binds `"127.0.0.1:0"` — the OS picks a free port. Returns `SocketAddr`
-/// (not bare `u16`) so callers can format `"http://{addr}"` directly via
-/// `SocketAddr`'s `Display` impl.
-///
-/// Per `adr-kikan-binary-topology §7`: desktop binds `:0` (ephemeral, never
-/// a fixed port); `mokumo-server` keeps the 6565–6575 range (`try_bind` above).
-pub async fn try_bind_ephemeral_loopback()
--> Result<(tokio::net::TcpListener, std::net::SocketAddr), std::io::Error> {
-    let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
-    let addr = listener.local_addr()?;
-    tracing::info!("Ephemeral loopback: listening on {addr}");
-    Ok((listener, addr))
-}
+// `try_bind_ephemeral_loopback` was lifted to `kikan-tauri::net` in S4.2 (#507).
+// The ephemeral-loopback bind strategy is specific to the Tauri shell and belongs
+// in the Tauri adapter crate (I2 hygiene). services/api deliberately does NOT
+// re-export from kikan-tauri — doing so would drag `tauri` into `mokumo-server`
+// via the mokumo-api dep, violating I3 (headless binary has no Tauri deps).
+// Desktop callers import `kikan_tauri::try_bind_ephemeral_loopback` directly.
 
 /// Generate a random setup token (UUID v4).
 pub fn generate_setup_token() -> String {
